+++ backend/core/src/auth/types/user-filter-keys.ts
@@ -1,0 +1,3 @@
+export enum UserFilterKeys {
+  isPartner = "isPartner",
+}